Add common definitions for testcase and command
This script adds the shell functions command() and testcase(). Both
check, if they are running in a LAVA testshell and execute LAVA specific
commands for test reporting. With this all shell scripts that use this
functions can be run either by a user or in a testshell.

Shameless copied from
https://git.lavasoftware.org/lava/functional-tests/-/blob/master/testdefs/lava-common

Signed-off-by: Frank Erdrich <f.erdrich@kunbus.com>