Add deployment job to write images onto the device
The script together with the YAML file are giving LAVA the ability to
program a new image onto a RevPi. A new image must be available in the
path lava-lxc, which is the default path for downloaded artifacts that
are made available for the used LXC container. LAVA controls all those
stuff, so there is no need to do anything by hand.

The script needs a RevPi in programming mode and also needs the location
of the USB device. The surrounding scripts and jobs should give the
information and do all necessary stuff to bring the device into
programming mode.

Signed-off-by: Frank Erdrich <f.erdrich@kunbus.com>