feat: Add first standard health-checks
A health check is a special type of test job, designed to validate that
the test device and the infrastructure around it are suitable for
running LAVA tests.

Health checks jobs are run periodically to check for equipment and/or
infrastructure failures that may have happened. If a health check
fails for any device, that device will be automatically taken offline.
Reports are available which show these failures and track the general
health of the lab.

The files health-check_RevPi_Core.yaml and
health-check_RevPi_Connect.yaml are the first standard health checks
used with RevPi devices.

health-checks files are placed in the server:
/etc/lavar-server/dispatcher-config/health-checks/

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>