feat: Create job to flash and test image with LAVA
"prepare_lava_job" is a tool to generate LAVA Jobs automatically.
Jobs are generated to flash a RevPi device and then perform a series of
tests, which are defined in a yaml file and passed as a parameter.

The generated Job has some variables already defined by default, which
can be modified using parameters. In the default behavior, the
generated Job is ready to be executed in a RevPi_Core and in our
Testrack (device_type: RevPi_Core, tags: config_core_01).

Once the LAVA Job has been generated, it could be submitted with can be submitted with
lavacli, webUI or REST/XML-RPC API.

For more information on how to use this tool, please see the
README.md file.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>