feat: Add job to flash RevPi_Core
In commit f96bb4e a test for writing images into the device was added.
The job submitted to LAVA in order to run this test is added with this
commit.
The image to be written was uploaded to JFrog and must be compressed in
xz format. In this case the image is
"2023-06-26-revpi-bullseye-rc3-arm64-lite.img", and thus can be used
for testing before our bullseye release.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>