docs: Add comments to flashing job
At first glance the flashing job for the RevPi_Core looks like a total
mess, but of course there is a lot of thought put into it that might
not be visible.
This aims to improve not only the readability of the flashing job but
also make it easier to comprehend what is happening and what
interactions with different parts of the system and configuration files
there are.

Additionally, links to the documentation are added where needed to
encourage looking at the (admittedly very hard to read and understand)
LAVA documentation.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>