feat: Add test to install another Kernel
In order to test new kernels, a LAVA test is added to facilitate their
installation.

The parameters KERNEL_url_armhf and KERNEL_url_arm64 contain the URLs
for accessing the kernel in the artifactory.

After installation, the device under test (DUT) will reboot
automatically after 1 minute, if the `SKIP_REBOOT` variable is set to
"false", which is the default case.
If the `SKIP_REBOOT` variable is set to `true`, the reboot is skipped.

If the DUT performs a reboot, the Job will be marked as FAIL, as the
connection will be lost.
This should be handled by the tester in the same way as with RTC tests,
i.e. a new deploy, boot and test section should be generated and used
for subsequent tests.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>