feat: Automate RTC tests
RTC-1:
Set hardware clock to the date and time specified in `DATE_SET` and
verify the expected time.

RTC-2-a:
Disable NTP synchronization, set hardware clock to the specific date
and time provided by `DATE_SET` variable.
If `SKIP_REBOOT` variable is equal to "true", it skips the reboot.
Otherwise, a reboot of the DUT will occur within 1 minute.
After the reboot, the RTC status can be checked using the independent
test `RTC-2-b`.
Note that NTP will not be automatically re-enabled after a reboot.
This must be handled by the tester in the script execution environment.

RTC-2-b:
Verify the expected time. This check is usually used after reboot of
test RTC-2-a.
By default this test will run with the `SKIP_REBOOT` variable set to
"true" so at the end of the test the NTP is re-enabled again.

NOTE:
Usually if a reboot is planned (SKIP_REBOOT != "true"), it would mean
that the RTC-2-b test will be executed after a restart.
After a test execution, it is crucial to ensure that NTP is re-enabled
and started to guarantee that subsequent tests have the correct time
configuration.
Therefore, 'timedatectl set-ntp true' is executed, but only if the DUT
should not be restarted.
This ensures that synchronization with the Network Time Protocol (NTP)
is maintained to keep accurate time in subsequent tests.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>