feat: Add notifications block in job
Users can receive notifications containing information about submitted
test jobs and the accompanying results by adding a notify block to
their test job submission.

For this purpose the user "automator" is used, which is also used in
the "lava-ci" repository. This user has an email account attached to
it, where the results of the job will be sent after the job is
finished. In our case, this email will be sent to an MS-Teams channel.

An email configuration located in
/etc/lava-server/settings.d/01-mail.yaml has been added.

For more information please see:
- https://docs.lavasoftware.org/lava/user-notifications.html#user-notifications-in-lava
- https://docs.lavasoftware.org/lava/user-notifications.html#direct-listing-of-recipients
- https://docs.lavasoftware.org/lava/data-export.html#notification-summary

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>