feat: Add security mirror in lava-lxc
After preparing the LAVA-Workers in RevPi-Connect, the security mirror
used by lxc-debian must be delivered as a parameter when using
lxc-create.
If the parameter is not given, the default mirror used is the
following:
deb http://security.debian.org/debian-security bullseye/updates main

This leads to an error in the container and the "apt-get update"
command cannot be executed, resulting in a LAVA-Job Error.

With this modification, the security mirror is previously defined in
the job:
http://security.debian.org/debian-security bullseye-security

Now the "apt-get update" command is executed correctly and the created
lxc container contains the necessary packages to run a LAVA-Job
correctly.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>