fix: Set Job timeout to 60 minutes
The `prepare_lava_job` tool uses the file `job-flash-RevPi_Core.yaml`
as a manifest to generate a job where a DUT is flashed and a testsuite
is executed.
If the testsuite used is long, a timeout occurs after 30 minutes.
This is solved with this commit, giving the Job generated by the tool
a time of 60 minutes before a timeout occurs.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>