fix(piControl): Install packages with apt directly
When installing dependencies for tests we're using sh-test-lib's
dist_name function which checks several files to find out what
distribution the tests are running on to be able to install test
dependencies for that specific distribution.

This doesn't work for our 32-bit image because the value returned there
isn't recognized by sh-test-lib's install_deps function.
Instead, install the dependencies with apt directly as our tests are
only running on debian systems, anyway.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>