refactor(eth): Update tests to new structure
In order to improve the eth tests, these tests are incorporated while
maintaining the defined structure.
The eth-1 test was implemented according to the previous test
`test-eth-ethtool.sh`.
The test eth-3 was implemented according to the previous test
`test-eth-iperf3.sh`.

The eth-1 test retains its original form, where it checks for the
existence of certain parameters after executing the `ethtool` command.
The eth-3 test has been slightly modified to check the performance of
the interface. The check is performed using the output of the `iperf3`
command in JSON format.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>