feat: Add script to generate test protocols
This script utilizes the LAVA API to extract test data and generate
protocols in Markdown format.

It takes the job number as an argument and generates a Markdown file
with information about test names, test cases, test IDs, and results.

It works by querying the suites and associated tests with the given job
number and uses LAVA result links to generate links in the Markdown
document.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>