feat(piTest): Add digin and relay tests - Connect*
To be successful with these tests, one leg of the relay must be
connected to +Vin and the other leg to the digital input of the
RevPi_Connect*.

The "RevPi_Connect_4" controls the relay through bit 0 of the
RevPiOutput variable.
Other "RevPi_Connect" controls the relay through bit 6 of the RevPiLED
variable.
The status of the digital input is always read with bit 6 of the
RevPiStatus variable.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>