feat: Add Bluetooth tests
This commit adds a new script for testing Bluetooth functionality.
At the moment the script includes three test cases:
- `bt-1`: Performs a Bluetooth device scan using `hcitool`.
- `bt-2`: Scans for a specific Bluetooth device using `bluetoothctl`
  and pairs with it.
- `bt-remove`: Unpair bluetooth device.

The `BT_REMOTE` parameter contains the name of the remote device to which
the RevPi will connect.
If this parameter is not provided, the script will display usage
instructions and exit.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>