feat(prepare_lava_job): Don't notify by default
Currently notifications are sent for every job that is started with
prepare_lava_job. This can cause notification fatigue and it's not clear
which jobs are important or not.
Add a flag "-n" or "--notify" that actually notifies the recipient. By
default, no notification is sent out, this setting is opt-in.

A use-case for this might be scheduled nightly builds that should notify
on incomplete jobs instead of reporting the status of every job,
regardless of whether it actually finished or not.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>