fix: Remove broken RTC-2 test from test suites
For unknown reasons the RTC-2 test, which tests if the hardware clock is
working, breaks LAVA. The test runs, makes its preparations in RTC-2a,
reboots, and can't be reached anymore. LAVA decides this is an
infrastructure error and takes the DUT offline. No more jobs will be
able to run on this DUT until its status is set back to "Good" manually.

This breaks the DUT until manual intervention after every test with
RTC-2. Avoid this by disabling the test for now until the reason for
this failure is researched and fixed.

Reported-by: Ramiro Omar Gsponer <r.gsponer@kunbus.com>
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>