fix(piControl): keep same test name for pass/fail
There is a test called "pc-2" and it can either fail or pass. Enforce
this.

"run_test_case" is used instead of "check_return" because the latter
returns the same return value as the command before it. This causes
problems if "check_return" is the last statement in the script which
makes the script return a value > 0. This, in turn, makes the plan fail
and not execute the "send-to-lava.sh" script which collects the results
of the test.
Because of this all results are discarded.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>