feat: Add wlan tests with 2.4GHz/5GHz
The AP has been modified to be able to perform 2.4GHz and 5GHz tests.

If the variable WLAN_24G or WLAN_5G is defined in the job definition,
the WLAN_SSID and WLAN_PASSWORD of this network will be used for the
test.
If the WLAN_24G or WLAN_5G variables are not defined in the Job, a
default WLAN_SSID network defined for the test will be used.

The variables WLAN_SSID_24G, WLAN_SSID_5G, WLAN_PASSWORD_24G and
WLAN_PASSWORD_5G must be previously defined in the device dictionary of
each DUT.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>