feat(rs485): make test more robust
The RS485 test is very brittle: If a message is lost then a timeout
*has* to be thrown from the test caller and the services for the RS485
test explicitly have to be restarted.

Additionally, the split into transmit and receive don't make sense here
as both tests both transmit and receive, making the split completely
arbitrary.

Reimplement the RS485 test in Python and make it more robust by making
the payload random and using a CRC to validate the data was
sent/received correctly.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>