fix(prepare_lava_job): allow multiple pkgs
The pkg-install.sh script to install packages as part of a test allows
multiple packages to be installed. The prepare_lava_job script doesn't
handle this correctly and only ever takes the last package specified to
be installed.

Fix this by allowing multiple --pkg-url arguments to be given which are
passed to pkg-install.sh.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>