fix(plans): don't deploy on testsuite-II
testsuite-II is the testsuite used for the second part of the RTC-2
test. testsuite ends with rtc-2a which prepares the date of the system
with a date in the past after which the system reboots. testsuite-II
then checks if the date set in rtc-2a is still correct in the test
rtc-2b. After the reboot and before connecting to the DUT again the plan
specifies to deploy the tests to the DUT *again*.

The deploy action causes to deploy a tar archive to the DUT where it is
unpacked. Since the tar archive is prepared on the worker that has a
normal date, this is what is embedded in the tar file. Unpacking this on
the DUT where the date is set in the past will give errors like

    tar: lava-8667/0/lava-test-runner.conf: time stamp 2024-11-04 03:34:26.2886791 is 5117270.291336578 s in the future

Deploying the tests onto the DUT was already done before the tests were
started and nothing about the tests changed since then, so just skip
this redundant deployment onto the DUT which avoids the above errors.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>