refactor: remove unnecessary install check
Since commit ceb5a1d8 ("feat: install dependencies with install_deps")
the "install_deps" function from sh-test-lib is used to install
dependencies. This function takes a boolean value as a string which
decides whether to actually install dependencies or not.

Remove the custom logic for installing dependencies and use the existing
logic instead.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>