refactor: Use symlink for usb_disk detection
Replaced the 'find' command used to dynamically locate the USB disk
with a static assignment to "/dev/blockDUT".

This change aligns with recent updates to the Worker configuration.

Workers now use `rpiboot -d mass storage gadget64` instead of the
previous mass storage gadget command. When running 'rpiboot', both the
primary device and USB devices connected to the DUT are detected.

Additionally, a new udev rule on the workers creates a symlink at
'/dev/blockDUT', which points to the correct Raspberry Pi module that
needs to be flashed. This ensures reliability as the module is not
always loaded as 'sda'. This is the device attached to the LXC
container and used for flashing.

The symlink '/dev/blockDUT' should now be used for consistent device
identification, removing the need for dynamic searches under
'/sys/devices'.

Signed-off-by: Ramiro Gsponer <r.gsponer@kunbus.com>