fix(hdmi): skip same test case
A test should always contain the same test cases. If a test is skipped,
the test case should reflect that.

The "start-desktop_check-default-image" test isn't really a test, it's
just a check if the actual test "start-desktop" can run. Remove the
"start-desktop_check-default-image" "test" and make sure the test case
"start-desktop" is skipped instead of the
"start-desktop_check-default-image" case.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>