feat(rs485): add rs485-dev-present test
Instead of outright failing the test and not listing other tests when
the RS485 device isn't present, make it a test itself and explicitly
skip the other tests if the device is missing.

The rs485-dev-present test shouldn't be added manually by the caller. It
will instead always be added as a test first because it is a
prerequisite for all the other tests.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>