fix(eth): always return 0 from run func
The run function always needs to return 0. The return value of a
function if no explicit return value is given is always the exit code of
the last command. If this command failed, the run function will fail,
which in turn makes the script fail, which would suppress proper test
output for LAVA. Instead make sure that the run function always returns
0 so that this doesn't happen.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>